Clean up and reorganisations after metadata download removals
