+++ golem-api-grpc/build.rs
@@ -43,7 +43,6 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
                 "proto/golem/common/project_id.proto",
                 "proto/golem/common/empty.proto",
                 "proto/golem/common/error_body.proto",
-                "proto/golem/common/json_value.proto",
                 "proto/golem/common/resource_limits.proto",
                 "proto/golem/common/uuid.proto",
                 "proto/golem/component/producer_field.proto",
