--- golem-api-grpc/proto/golem/common/json_value.proto
@@ -1,36 +1,0 @@
-syntax = "proto3";
-
-package golem.common;
-
-message JsonValue {
-  oneof kind {
-    NullValue null_value = 1;
-    PosIntValue pos_int_value = 2;
-    NegIntValue neg_int_value = 3;
-    double number_value = 4;
-    string string_value = 5;
-    bool bool_value = 6;
-    JsonArray array_value = 7;
-    JsonObject object_value = 8;
-  }
-}
-
-message JsonArray {
-  repeated JsonValue values = 1;
-}
-
-message JsonObject {
-  map<string, JsonValue> fields = 1;
-}
-
-message PosIntValue {
-  uint64 value = 1;
-}
-
-message NegIntValue {
-  int64 value = 1;
-}
-
-enum NullValue {
-  NULL_VALUE = 0;
-}
