+++ golem-service-base/src/lib.rs
@@ -15,10 +15,8 @@ pub mod api_tags;
 pub mod auth;
 pub mod config;
 pub mod db;
-pub mod exports;
 pub mod model;
 pub mod routing_table;
 pub mod service;
 pub mod stream;
 pub mod type_inference;
-pub mod typechecker;
