+++ golem-worker-service-base/src/service/component/default.rs
@@ -198,9 +198,9 @@ pub struct ComponentServiceNoop {}
 
 impl ComponentServiceNoop {
     pub fn test_component() -> Component {
+        use golem_common::component_metadata::ComponentMetadata;
         use golem_service_base::model::{
-            ComponentMetadata, ComponentName, ProtectedComponentId, UserComponentId,
-            VersionedComponentId,
+            ComponentName, ProtectedComponentId, UserComponentId, VersionedComponentId,
         };
 
         let id = VersionedComponentId {
